External versions: save state (open / closed)
When an external Version is created, we mark it as `.state="open"` and when it's
closed/merged we mark is as `.state="closed"`.

Note that we cannot use `.active=False` for this as we were doing, because we
want to keep serving closed/merged PRs during some extra time after they are
closed/merged and El Proxito is not serving inactive versions.

With this approach, external versions will be always `.active=True` but with
different `.state=`. That state is used by the periodic Celery task to know if
the version has to be deleted or not.

Closes #9109