Build: POC of `build.commands`
Minimal implementation of `build.commands` as a POC to show how it could work.

It's using a `.readthedocs.yaml` similar to this one:

```yaml
version: 2

build:
  os: ubuntu-20.04
  commands:
    - mkdir output/
    - echo "Hello world" > output/index.html

  tools:
    python: "3"
```

This, of course, is not a good implementation and it's done pretty quick as a
way to show what parts of the code are required to be touched. I think this will
help with the discussion about how the UX around `build.commands` could work.

It defines an "implicit contract" where the `output/` folder under the
repository checkout will be uploaded to S3 and no HTML integrations will be done.