Spruce up docs on pull request builds
I keep helping users with the support request around webhooks missing
the pull request event, so wanted to clarify a bit here. I also took the
opportunity to add some clearer language to the rest of the page, and
resturcture a few things to improve the information flow.

I also opened #9176 to deal with this in a programmatic way, but figured
we could at least communicate the issue better for now.