Remote repository: Keep in sync when repos are moved to another org/user.
We were doing some checks to avoid removing the remote organization
from a remote repo when this is listed from the /users/repos/ endpoint
or changing the organization when the project was moved.

But, since we are hitting the GitHub API directly,
we can always trust the result from there,
which includes the type of owner of the repository,
we can use that to decide if the repository belongs to an organization or not.

Fixes https://github.com/readthedocs/readthedocs.org/issues/8715