Fix copy issues on model fields (#9170)
* Fix copy issues on model fields

The term "domain" is not a proper noun and should not use proper
capitalization. We do this in a number of places, as we are generally
referring to the underlying model, but this shouldn't be how these terms
are communicated to users.

I didn't touch the actual copy otherwise, even though we could probably
make these descriptions a bit clearer. We could also do that more
immediately if it makes sense.

* Add migration

* Reformat