Assets: update `package-lock.json` with newer versions
While doing the release, `npm ci` hangs when trying to build the assets because
it doesn't find some outdated packages that we have in our `package-lock.json`.

I removed the `package-lock.json` file and build the assets with `inv
docker.buildassets`, which generates an updated `package-lock.json` using the
most recent packages.

After that, I did a quick QA locally mainly in the build detail's page to check
if the Javascript that polls the status works and it worked. However, we will
need to QA some of the Javascript aspects after the deploy as well.

Hopefully, everything works.