Build: update the `Version.documentation_type` properly
When updating the `Version` object via the API on success build, we were using
the same value we received from the API for the `Version` object.

This commit uses the `doctype` value from the config file used to build the
version. This way, we keep the `Version.documentation_type` updated when there
are changes in the config file for a particular version.