Subscriptions: handle subscriptions with multiple products/plans/items
subcription.plan is deprecated in favor of subscription.items,
this is only a problem for subscriptions that have more than one product
(subscriptions.plan will be null for those cases).

On .com have only a couple of subscriptions with multiple plans/products,
the features they provide are being added manually at the moment,
so we don't need to track those in our DB yet.

Ref https://sentry.io/organizations/read-the-docs/issues/3048896842/?environment=production&project=161479&query=stripe&statsPeriod=14d