EmbedAPI: handle special case for Sphinx manual references
When using `.. _my-reference:` above a title, Sphinx adds an empty `<span
id="my-reference"></span>` to the resulting HTML.

Then, our backend is queried for `my-reference` fragment and as it finds it
immediately, it returns the corresponding HTML tag. The problem is that tag is
empty, so the tooltip shows no content at all.

This commit handles this special case when the API is called with
`doctool=sphinx` and returns the parent node of this `span` tag which is the
section itself.

See https://github.com/readthedocs/sphinx-hoverxref/issues/199 for a longer
description of the problem.