Local development: use `nodemon` to watch files instead of `watchmedo` (#9338)
* Local development: use `nodemon` to watch files instead of `watchmedo`

We have been dealing with a problem with `watchmedo` that restart the process
when it shouldn't be restarted and it doesn't restart it when it should.

I got tired debugging `watchmedo` and I was suggested to give it a try to
`nodemon` because we are already using it in other projects. I'm not super happy
adding a node dependency to the Dockerfile, but I didn't find a better way to do it.

I did some tests with the configuration proposed and it seemed to work pretty
well. I'm sure we will find some edge cases, but we always can tune it a little
more later.

Closes https://github.com/readthedocs/readthedocs.org/issues/8802

* Local development: remove `watchdo` Python package

It's not required anymore since we are using `nodemon` now.