Development: avoid path collision when running multiple builders
When starting up the local development with multiple builders (e.g. `inv
docker.up --scale-build=2`, or more) it's possible to find them collisioning on
the `user_builds/` path when building the same version on each builder.

This happens locally but not on production, because we are running all of them
on the same host machine using the same disk/volume for all of them.

To avoid this collision, this commit adds an extra directory with the `hostname`
of the container to the `DOCROOT` to avoid this collision.