OAuth: connect `Project` to `RemoteRepository`
There are multiple reasons why a `Project` could be disconnected from a
`RemoteRepository`:

1. it was imported manually
2. the user deleted their account, which also deleted the
   `RemoteRepository` (old behaviour)
3. got disconnected when doing the data migration
4. others?

If any of the previous scenarios happened, the user was not able to (re)connect
their project anymore. The only way was to delete the project and re-import it
using the wizard with a VCS account connected.

This commit adds a button under "Admin" -> "Integrations" page to "Connect VCS
repository" allowing users to perform this action by themselves.

When a `Project` is connected to a `RemoteRepository` permissions can be managed
at VCS level, allowing the organization to be able to use VCS SSO. Besides, this
connection will be used when sending status notifications to the VCS.