Build: handle `on_timeout` properly
We were accessing to `self.task.data` which does not exist. `self.task` is not
the same instance as the one it's running, so we don't have `.data` on it.

However, we can access the arguments passed when called. So, we are logging that
information which is what we have.

This commit solves a problem reported on Sentry everytime this happens.