Build: refactor TaskData
Currently this class holds any values,
defining its attributes dynamically,
this makes it hard to know what attributes
exactly does this object hold,
and others are initialized in several parts of
the codebase.

So, I'm defining all the attributes in a form of a dataclass,
with `slots=True` (so no other attributes can be added by mistake).

I'm also checking for possible uninitialized data in the task handlers.

Note: Dataclasses require type annotations,
this doesn't mean we are using type hints or enforcing them in our codebase.

Closes https://github.com/readthedocs/readthedocs.org/issues/9364