Analytics: make sure there is only one record with version=None
By default Postgres doesn't consider two null values equal,
so you can create a duplicated record with a null value,
even if it has a unique constraint.

Postgres does provide a way to change that (`NULLS NOT DISTINCT`).

https://www.postgresql.org/docs/15/ddl-constraints.html#DDL-CONSTRAINTS-UNIQUE-CONSTRAINTS

But to use it from Django, we would need to execute a raw sql
expression.

So, another approach to avoid executing a raw sql is to create two constraints,
one with the normal constraint of unique fields,
and another one that doesn't include the optional field,
but it includes a condition (version=None).

Note that we are using get_or_create,
which does properly check if there is already a version
with `None`, but if more than on request are executed at the same time,
then get_or_create will end up inserting the record more than once
since the unique constraint with `None` wasn't enforced at the db level.

Closes https://github.com/readthedocs/readthedocs.org/issues/9316