Docs: improve the flyout page to include a full example
- add the Languages section
- add the Downloads section
- include some HTML comments explaining when the sections are present

Note that this is _not a real example_ since Flask does not have translations
nor downloads, but I don't think that's super important for the purpose of the
example. The most important is the structure of the HTML than the URLs included
on it.

We could use a real example if we find a good one, tho.