Redirects: merge query params from the redirect and original request
We were overriding the query parameters
instead of adding them.

I'm adding the ones from the redirect at the end
(should take precedence on some frameworks).
We could get fancy and replace them instead of adding
them, but it gets tricky when we have an array...
So I think this implementation should be fine.

Closes https://github.com/readthedocs/readthedocs.org/issues/9401