Redirects: check only for hostname and path for infinite redirects
We were comparing the proxito_path
which doesn't include query parameters with the final URL which does
include them, causing an infinite redirect.

This same thing could happen if someone writes a redirect to
`http://`, we redirect to the https version, then the redirect will
redirect to the http version again.