Build: unpin Pillow for unsupported Python versions
Pillow supports specific versions of Python. They are shown in this table:
https://pillow.readthedocs.io/en/stable/installation.html#python-support

For unsupported Python versions, they don't offer wheels. Meaning that Read the
Docs needs to compile the Pillow version on each build.

Ideally, we should unpin Pillow completely and install always the latest version
that's supported for this particular Python (as we do with other requirements
like Sphinx or MkDocs).

However, this commit kept old Python versions to keep using `Pillow==5.4.1`.