Dependencies: use GitHub Action + pip-tools
I finally decided to use a custom GitHub Action + pip-tools because:

- pyup does not support pip-tools
- pur does not pin transitive dependencies
- dependabot says it supports pip-compile but doesn't work

So, we are now running `pip-compile` weekly and creating a Pull Request with the
changes performed to the output files. From there, we can decide if we merge it
or not.

Note the GitHub action handles the creation of the PR (or the skip of it)
automatically and should not flood us with multiple PRs.

Also note that I'm doing this for `requirements/docs.in` in this PR only to show
how it works and avoid the noise, but I will prepare another PR that uses the
same pattern for all the other requirements files.

Closes #9273