Telemetry: collect user's requirements
We are saving `packages.pip.user` that are "top level packages" --which does not
mean anything too useful for our purposes.

The original idea was to have `packages.pip.all` that lists all the packages
installed in the virtual environment and `packages.pip.user` that lists all the
packages installed because the user specified them in the requirements file.

This PR parses all requirements files specified by the user and saves package
names and versions into `packages.pip.user`. These ones are pretty useful to
know what projects are specifically defining/pinning a package.