Code comment (#9518)
Just a comment saying that the function does not work if the `request=` argument
is not passed. We should fix this somehow.

Note that this is not currently failing because we are not calling this function
without the `request` argument. However, it should be mandatory or it should
support not passing it properly.