Query: minor improvement
Limit the builds to those that are from the last day (we won't have a build
length that's greater than 1 day). This will reduce the rows required to parse
since we have an index there. Also, remove the "equal" part from "less than
equal" from the query because it makes no sense.

This may help with some of the db problems we were facing in last weeks.