Build: skip build based on commands' exit codes
Define a particular exit code (439) to skip a build.

If any of the commands returns this exit code, the build will be cancelled
automatically and won't run any of the following commands. When this happens,
the build will be marked as `cancelled` and no email/webhook notifications will
be sent.

Why 439 was chosen for the exit code?

It's the word "skip" encoded in ASCII. Then it's taken the 256 modulo of it because
the Unix implementation does this automatically for exit codes greater than 255

```
>>> sum(list('skip'.encode('ascii')))
439
>>> 439 % 256
183
```