Clean up some old code that's not used
I found some of these methods confusing when writing new code. One of them
seemed to do what I wanted, but it was using old tricks (find a `conf.py`
instead of using `sphinx.configuration` key from the config file).

I think it's a good idea to remove them to avoid mistakes in the future. It
seems they are not used anywhere and it's safe to remove them.