ImportedFile: use BigAutoField for primary key
We could disable search indexing while we do the migration,
but I don't think that should be required, we have 11M records,
but to migrate the SphinxDomain model it took 15 min,
and we had ~56M.

```python
In [7]: ImportedFile.objects.count()
Out[7]: 11527437
```

So some 3 min of not being able to index new versions doesn't seem
bad... There are two things that could happen:

- The query times out and we don't index that version.
- The query waits till the migration is done,
  nothing gets lost.

But if we disable search indexing we definitely
won't index new versions.

We don't use those models outside search indexing,
so doc serving and such shouldn't be affected.

ref https://github.com/readthedocs/readthedocs.org/issues/9492