Admin: install `debug_toolbar`
This application needs to be installed if we want to enable it later in the
`web-extra` instance. Otherwise, it fails due the static files are not in the
manifest.

By installing it by default, `collectstatic` will find this application and copy
these static files and add them in the manifest.

Note this application will only be used if `SHOW_DEBUG_TOOLBAR=True`, which is
`False` by default.