Merge pull request from GHSA-368m-86q9-m99w
* Prevent symlink following in path operations

* Black

* Fix tests

* Updates

* Tests

* Add tests

* Tests

* Updates from review

* Use build exception

* Typo

* Fix tests

* Fix tests

* Always raise an exception if the usage of symlink is invalid

- the symlink points to a file that does not exist
- the symlink points to a file _outside_ the DOCROOT/base_path
- symlinks are not allowed in the operation

In any of these cases, we hard fail the build and communicate a generic
error (for now) to the user: "Symlinks are not fully supported". We can improve
the message with more value in another commit/iteration.

These are 3 different exceptions, so we can get as much granular as we want
here. All of them are managed by Celery `on_failure` handler, which will
automatically do the correct thing to finalize the build properly.

* Convert `path=` to str() to avoid "not JSON serializable" issue

* Update tests to check for new exceptions instead of `None`

* Run `pre-commit` since it was not enabled by default

Co-authored-by: Manuel Kaufmann <humitos@gmail.com>