Build: handle symlinks in a centralized way
Simplifies symlinks hanlding to centralize the security logic in one place
only (`safe_open`); reducing the complexity of this code, making it easier to
maintain and more secure.

Besides, it improves the UX for users to understand why their build is failing.
It clearly communicates the error using our internal pattern for known exception
managed at `Celery.on_failure`.

These are the highlights of this work:

- Always follow symlinks
- Always copy symlink as-is. Only resolve them when uploading to storage.
- Always check symlinks are inside DOCROOT and `base_path` if passed
- Hard fail the build if there are symlinks in the output that are outside
  DOCROOT
- Centralize all the symlink checks in `safe_open` (used to open configuration
  files while building and to open files for uploading them)
- Clearly communicate the error to the user
- Move upload to storage function inside `Celery.execute` handler to allow us
  raising `UnsupportedSymlinkFileError` and using `on_failure` to clearly
  communicate the error to the user