Build details page: normalize/trim command paths (#9815)
* Build details page: normalize/trim command paths

Small hack to trim/normalize command paths when displaying them to the user
under the build detail's page.

This is a simple hack and there are more we can do here. Probably, there are
some edge cases this is not catching. However, the worst it can happens is the
path not being trimmed.

* Make regex to work locally and in production

* pre-commit