EmbedAPI: decode filepath before open them from S3 storage
We can use encoded URLs (e.g. with %20 for white spaces) for external requests
since they are valid URLs. However, when it's internal and we get it from the
storage, we have to convert them to regular characters (white spaces, in this
example).

We use `urllib.parse.unquote` for this before passing the filename to the S3
backend storage.

Closes #8301