CORS: don't allow to pass credentials by default
Previously, we were allowing to pass cookies in cross-site requests
(because our sustainability API required it), but due to this,
we are doing some checks to guarantee that these requests weren't
accessing to private resources.
These checks are getting a little complex (they are at the middleware
level), since there are several things that influence if the resource is
public or private.

This doesn't represent a breaking change since:

- The endpoints listed bellow allow to
  access public resources only.
- There is no need to pass cookies for
  accessing public resources.
- Cookies were only allowed to be passed in cross site requests on .org.
- For .com our session cookies are set to `Lax`,
  so they were never included in cross-site requests.

This could break if users were passing the `{'credentials': 'include'}` in their requests
(browsers will reject those requests now), they just need to remove that from their code,
or use the proxied APIs. If they were trying to access private resources, that wouldn't have
worked anyway.

\## Notes

I'm not changing the current list of allowed endpoints yet,
wan to test that this works in production first,
but it worked locally.
Once we are settled, we can remove all the custom logic from
core/signals.py and just keep a list of URLs.

\## Endpoints allowed in cross-site requests:

- /api/v2/footer_html
- /api/v2/search
- /api/v2/docsearch
- /api/v2/embed
- /api/v3/embed