Build: check if the output directory is a directory
It checks for `_readthedocs/<format>` output directory to be a directory.
If it's not a directory, it silently skip that format for now.
This behavior will change once we move out `store_build_artifacts` from
`on_success` and we can communicate errors easily to the user.

Related https://github.com/readthedocs/readthedocs.org/issues/9931