MkDocs builder: use proper relative path for `--site-dir`
After performing the standardization of output directories,
we broke MkDocs builders for those projects that don't have the config file in
the root of the repository.

This happens because MkDocs consider the output directory relative to where the
`mkdocs.yml` file is, instead from where the command is executed.

This commit solves this problem by calculating the `--site-dir` relative to
where `--config-file` is.

Closes #9935