Settings: simplify all the settings removing a whole old layer (`dev`) (#9978)
* Settings: simplify all the settings removing a whole old layer (`dev`)

We are not using `settings/dev.py` since we started using
Docker (`settings/docker_compose.py`).
However, we never updated this code.

This commit removes the extra `Dev` class layer by moving the required (and
still used settings) into `settings/test.py` and `settings/docker_compose.py`
since they were inheriting from `settings/dev.py`.

* Settings: add missing setting for test

* Settings: missing on web

* Setting: missing on celery

* Settings: missing on proxito

* Settings: updates for docs/

* Docs: remove outdated and complex code and dependencies (#9981)

* Docs: remove outdated and complex code

This commit reduces the complexity of the documentation environment by removing
the requirement of having Django installed and loaded to build the docs.

* Remove the `djangodocs` and `doc_extensions`: they load the whole Django
application unnecesarily. We need to have a bunch of requirements installed just
because of this.

* Remove defaults from "Interesting settings" because all the default were
wrong. The Django extension was loading an invalid set of settings. This page
could be completely removed as well probably.

* Remove `featureflag` role because of the same reason (loads a whole Django
project). We were using it only to show a small description of the feature
flags. I wrote that text directly in the page. Feature flags are deprecated and
this page will probably deleted soon as well.

* Remove `buildpyversions` role also. Same reason. Manually wrote the versions
supported on each Docker image. This is used for the legacy images, tho. They
won't change since we don't maintain them anymore.

* Docs: reduce requirements drastically

Since we don't depend on Django anymore we can remove the `-r pip.txt`
dependency from the documentation.

This makes the documentation just to rely on Sphinx and some other extensions
only simplifying the environment a lot and running fast.

* Docs: don't install `pip.txt`

* Docs: remove missing `djangosetting` role

* Solve minimal merge conflicts