Build: organize absolute output directory for host and container
There are some commands that are executed from inside the container where
`$READTHEDOCS_OUTPUT` variable is defined and we can use it.

However, there are other commands executed from the host where that variable is
not defined and/or it cannot be used (e.g. as a `cwd=` argument for Docker run
command).