Celery: cleanup `pidbox` keys
Simple task to remove `pidbox` keys older than 15 minutes.
This is a workaround to avoid Redis OOM for now.

We will need to find out a better solution here. There is an upstream issue
opened that we should check in the near future and probably remove this
workaround: https://github.com/celery/celery/issues/6089