Proxito: simplify caching logic
Caching is enabled for all projects now.
We always check for the privacy level to explicitly cache or
not a response, if the response isn't explicitly cached or not,
we use the default, this default depends if we are on .org or .com
(we cache everything on .org, and on .com we cache responses attached to
public versions).

Closes #10031