Proxito: use `django-cacheops` to cache some querysets (#10075)
First attempt to use `django-cacheops` on El Proxito to cache some frequenly
used queryset when serving docs via El Proxito. Queries to these models,

- Project
- Version
- ProjectRelationship
- Domain

are done on every request, and we want to do a small test by caching them in
Redis to take a look at the metrics and decide whether or not this is faster
than Postgres when serving documentation.