Celery: cheat `job_status` view to return `finished` after 5 polls
The frontend hits this endpoint every 1s. As we don't know when the task has
finished now since we disabled the Celery result backend, we are just cheating
it and return finished after 5 times the view is called with the same `task_id`.

This fixes the issue we have in the UI when clicking the circled arrows to
re-sync the remote repositories.