Downloadable artifacts: make PDF and ePub opt-in by default (#10115)
Newly created projects will have PDF and ePub disabled by default. This matches
the default value of `formats` in the YAML file v2.

Building PDF and ePub is not a trivial task and many projects start failing
because of this or, if succeeding, they are just building pretty low quality PDF
and ePub files --which is a waste of resources.

This commit disables this by default for new projects, keeping the old projects
working as they are currently.