Build: pass shell commands directly (`build.jobs` / `build.commands)` (#10119)
* Build: pass shell commands directly (`build.jobs` / `build.commands)`

I'm not sure about _why_ we split the commands before sending it out our
executor. The command is finally wrapper properly before being executed.

I did some locally QA after removing the split of the command and it seems these
scripts work fine. So, it seems safe to move forward with this, explanding the
flexibility of our config file.

Closes #10103

* Test: match changes

It compares to a full command instead of being split.