Build: fail PDF command (`latexmk`) if exit code != 0 (#10113)
* Build: fail PDF command (`latexmk`) if exit code != 0

We were forcing `exit_code=0` if we detected that `Output written on (.*?)` was
found in the standard output of the `latexmk` command. This was producing
confusions to our users because the PDF generation failed but we were
considering it successful.

Now, we are relying on the exit code itself as we do for all the other commands
making sure the command succeeded before proceding with the build. Now, if the PDF
command fails, we fail the build completely.

* Test: adapt tests to the changes on PDF building

* Build: specific message for PDF generation failure

* Better user message

Co-authored-by: Eric Holscher <25510+ericholscher@users.noreply.github.com>

---------

Co-authored-by: Eric Holscher <25510+ericholscher@users.noreply.github.com>