Search API: `@this` alias support
In order to make it easy to search on
projects without having to know the current slug of the project,
I'm adding support for the `@this` alias.

I'm also allowing cross site requests to this API, since it can be useful
(and makes it easier to test the search extension locally :D).

Other implementations I took into consideration:
- Get the current project/version from the `Referer` header.
  This would be using the unresolver, but it felt a little hidden.
- A separate `@this` for project and version, e.g `project:@this/@this`.
  Apart from ugly, don't think it was useful to search for a different version
  from the default. But we can support `project:@this/v2` in the future
  if needed.