Build: pass `PATH` environment variable to Docker container (#10133)
* Build: pass `PATH` environment variable to Docker container

Instead of prepending all the commands with the `PATH=` variable, we pass the
environment variable directly to the Docker container.

This allow us to run multi-line shell script without failing with weird syntax
errors. Besides, the implementation is cleaner since all the environment
variables are passed to the commands in the same way.

I added some _default paths_ that I found by checking the local Docker
container. I'm also passing the users' path, depending if we are working locally
as root or in production.

This is not 100% complete and there may be some other issues that I'm not seeing
yet, but I think it's a first step to behave in a way our users are expecting.

Closes #10103

* Lint: for some reason fails at CircleCI otherwise

Locally it tries to reverted back 🤷

* Update readthedocs/doc_builder/environments.py

Co-authored-by: Santos Gallegos <stsewd@proton.me>

* Increase code's readability by applying suggested changes

---------

Co-authored-by: Santos Gallegos <stsewd@proton.me>